storage: factor out replicaStateLoader
This is one of the self-contained threads to unravel for #18784.
