+++ pkg/storage/replica_state_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
+	"github.com/cockroachdb/cockroach/pkg/storage/stateloader"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
@@ -58,20 +59,20 @@ func TestSynthesizeHardState(t *testing.T) {
 		func() {
 			batch := eng.NewBatch()
 			defer batch.Close()
-			rsl := makeReplicaStateLoader(cluster.MakeTestingClusterSettings(), 1)
+			rsl := stateloader.Make(cluster.MakeTestingClusterSettings(), 1)
 
 			if test.OldHS != nil {
-				if err := rsl.setHardState(context.Background(), batch, *test.OldHS); err != nil {
+				if err := rsl.SetHardState(context.Background(), batch, *test.OldHS); err != nil {
 					t.Fatal(err)
 				}
 			}
 
-			oldHS, err := rsl.loadHardState(context.Background(), batch)
+			oldHS, err := rsl.LoadHardState(context.Background(), batch)
 			if err != nil {
 				t.Fatal(err)
 			}
 
-			err = rsl.synthesizeHardState(
+			err = rsl.SynthesizeHardState(
 				context.Background(), batch, oldHS, roachpb.RaftTruncatedState{Term: test.TruncTerm}, test.RaftAppliedIndex,
 			)
 			if !testutils.IsError(err, test.Err) {
@@ -81,7 +82,7 @@ func TestSynthesizeHardState(t *testing.T) {
 				return
 			}
 
-			hs, err := rsl.loadHardState(context.Background(), batch)
+			hs, err := rsl.LoadHardState(context.Background(), batch)
 			if err != nil {
 				t.Fatal(err)
 			}
