+++ pkg/storage/store.go
@@ -46,6 +46,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlutil"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
+	"github.com/cockroachdb/cockroach/pkg/storage/stateloader"
 	"github.com/cockroachdb/cockroach/pkg/storage/storagebase"
 	"github.com/cockroachdb/cockroach/pkg/util/bufalloc"
 	"github.com/cockroachdb/cockroach/pkg/util/envutil"
@@ -1837,8 +1838,8 @@ func splitPostApply(
 		// committed the split Batch (which included the initialization of the
 		// ReplicaState). This will synthesize and persist the correct lastIndex and
 		// HardState.
-		rsl := makeReplicaStateLoader(r.store.cfg.Settings, split.RightDesc.RangeID)
-		if err := rsl.synthesizeRaftState(ctx, r.store.Engine()); err != nil {
+		rsl := stateloader.Make(r.store.cfg.Settings, split.RightDesc.RangeID)
+		if err := rsl.SynthesizeRaftState(ctx, r.store.Engine()); err != nil {
 			log.Fatal(ctx, err)
 		}
 	}
@@ -3121,7 +3122,7 @@ func (s *Store) processRaftRequest(
 		needTombstone := r.mu.state.Desc.NextReplicaID != 0
 		r.mu.Unlock()
 
-		appliedIndex, _, err := r.raftMu.stateLoader.loadAppliedIndex(ctx, r.store.Engine())
+		appliedIndex, _, err := r.raftMu.stateLoader.LoadAppliedIndex(ctx, r.store.Engine())
 		if err != nil {
 			return roachpb.NewError(err)
 		}
